backend/requirements.txt
fastapi
uvicorn
motor
pymongo
python-jose[cryptography]
pyjwt
bcrypt
python-dotenv
pydantic[email]